Featurize agent
